Easy Windows installation
Although not strictly necessary install qgit in git
directory so to easy setup and starting.

Also teach qmake to generate a launching script to
set correct PATH before to start the application. It's
up to the user create a link on the desktop to this .bat
file so to make qgit running a double click away.

Update documentation to reflect changes.

Signed-off-by: Marco Costalba <mcostalba@gmail.com>