Delay cache loading a little bit and fix a bug
So to let the user to see the first revision selected and
with the files loaded directly from git.

Starting immediately the cache load could block the UI while
file list is still empty. So delay of 500ms.

Also do not forget to resetting 'cacheNeedsUpdate' before
to load to avoid spurious cache savings when exiting.

Signed-off-by: Marco Costalba <mcostalba@gmail.com>