Fix disappearing merge files list when browsing revs
In case reading of file list from git is canceled because
of a new update request arrives we stop git external
process, but remain in the update loop with wrong data.

Normally this is harmless because the new pending update will
cleanup everything, but with merge revisions, in rare case, a
stale and persistent file list is created.

This patch raises an exception in case a new update request arrives
while in updating, so to exit the loop as soon as possible without
changing state data.

Some modification to exception handling has been required to avoid
unuseful warning messages in these cases.

Signed-off-by: Marco Costalba <mcostalba@gmail.com>