qgit: Default the log window to wider than the file list and set monospaced font
The log window and changed files list would start at equal sizes.  This
patch gives them each a horizontal stretch factor making them have a 3:1
ratio.

The font for the log viewer is set to "Monospace", but this should
really be done in software and given a config setting (perhaps use the
patch view font)

Signed-off-by: Andy Parkins <andyparkins@gmail.com>
Signed-off-by: Marco Costalba <mcostalba@gmail.com>