Fix opening a repo from a read-only filesystem
In case our repo is on a read-only filesystem or simply
we don't have write access to repo working dir then an
error occurs when trying to opening the repository.

We try to save the script for redirecting git-rev-list to
temporary file in working dir. This fails and opening is aborted.

Fix is to save also the script in temporary directory, together
with our data file.

Signed-off-by: Marco Costalba <mcostalba@gmail.com>