Use temporary file as data exchange facility also under Windows
QProcess interface is real slow, so tweak stuff a little bit
to use temporary file as default also under Windows.

Of course in this case the file is on a ntfs (not tmpfs) filesystem.
Another issue is that reading a repository mounted under ntfs from
Linux side requires QProcess facility due to ntfs being read-only
(Linux side), so we cannot create the startup script used to run
'git rev-list'.

As documented, you need to tweak a define in dataloader.h
and recompile to change facility.

Signed-off-by: Marco Costalba <mcostalba@gmail.com>