Fix an annotation issue with file deletion/re-creation
When a given path is deleted in one revision and then
recreated in a subsequent revision then path history
is truncated to file deletion revision if 'git rev-list'
is called with '--remove-empty' option.

This could lead to misbehaviours and also segfaults in
some cases when reading the annotation of such a file.

Thanks to Pavel Roskin for reporting the bug.

Signed-off-by: Marco Costalba <mcostalba@gmail.com>