Fix a bad annotation issue with symlinks
Symlinks in git are one line files with just the name of the target,
not the target content.

We use 'cat' command to get file content if the current selected
revision is the faked working dir revision ZERO_SHA '0000...'

But 'cat' resolves symlinks and returns target content, and this
leads to a 'bad annotation' assert when trying to open file viewer.

With this patch we use 'cat' only if the file is modified in working
dir, to let annotation work for changed files still unknown to git,
otherwise we go with a safe 'git cat-file blob HEAD' instead.

NOTE: There is still an issue if the modified file is a new symlink,
just converted from an old plain file. In this case annotation will
fail until change is committed.

Signed-off-by: Marco Costalba <mcostalba@gmail.com>