Fix saving of a binary file
When using 'Save file as...' menu entry to save on disk a
tree file (blob) object a check on extension is done to identify
the file as image or archive and, in case, saving with row byte
content, without string encoding.

Note: Because git does not recognizes binary file content we use
file name extension to infere MIME type.

Signed-off-by: Marco Costalba <mcostalba@gmail.com>