Workaround limited stdin process buffer size in Windows
This patch fixes a very bad limitation under Windows platform.

Because process input pipe buffer is quite small on that platform
we have to use a temporary file and redirect process standard input
to that file.

This avoids crashes and hangs when, as example, annotating a file with
a long history (that is passed in input pipe).

Also remove an hack on the number of filenames to load at startup
now that is no more necessary.

Signed-off-by: Marco Costalba <mcostalba@gmail.com>