Use better icons for turning off added and removed lines
The current icons use too little details that are hard to see.  They are
also unfriendly to people with color vision defects.

Instead, use green plus and red minus with thicker lines and gradient.
When the plus or the minus is disabled, use a black-white pattern that
is visible on any background, not too bright and easy to distinguish
from solid colors.

Patch of Pavel Roskin ported from stable series.

Signed-off-by: Marco Costalba <mcostalba@gmail.com>