Optimize running git commands
Currently, after we start a git command we check pending events then
wait 20ms. If process is still running we redo the check+wait loop

Reverse the loop instructions to wait before to check, so that fast
commands that take less then 20ms to complete avoid to wait for
another useless loop cycle.

Signed-off-by: Marco Costalba <mcostalba@gmail.com>