Use 'git show-ref' to read StGIT patches sha's
These days StGIT stores patches sha as git references under refs/patches/

So use this, while reading the other refs, instead of walking
internal StGIT directory tree seeking for sha values.

When StGIT support was added patches where not mapped as
references, so we were forced to use this hacky StGIT
tree walking. We can now happily avoid to mess with
StGIT internals.

As a nice side effect a lot of code went away and we got a good
speedup too thanks to avoiding a tree walking.

NOTE: This patch could break compatibility with very old StGIT
archives, in that case an update of StGIT is recommended.

Signed-off-by: Marco Costalba <mcostalba@gmail.com>