Fix an issue when releasing "Pin view" check box in file viewer
After pinning down a file with 'Pin view' check box in file viewer,
then moving to a different revision in file history and then releasing
the check box a "bad annotation" assert could occur.

This is an old bug, introduced in Agust 2006 when support for
source highlighting was added!

It was hidden till now because it's not easy to trigger.

Signed-off-by: Marco Costalba <mcostalba@gmail.com>