Run 'git rev-list' with '--stdin' option
To avoid command line arguments exeeding platform limits use
'--stdin' option to pass arguments in an input buffer.

With this patch we should scale to StGIT archives with hundreds of
unapplied patches or standard archives with a lot of branches, that
are passed to 'git rev-list' when loading a file history.

Signed-off-by: Marco Costalba <mcostalba@gmail.com>