Convert loading to a complete async operation
Currently we wait on a loop while loading StGIT unapplied patches. This
is at least ugly and not recommended by Qt.

Convert to use a two step initialization, in the first step we start loading
unapplied patches and return immediately to event loop, at the end we load
remaining revisions.

As a side effect we don't need anymore the "wait till stopped" hack in
MainImpl::setRepository() because now Git::stop() is completely non blocking.

This is a 'difficult patch', a lot of test is needed.

Signed-off-by: Marco Costalba <mcostalba@gmail.com>