Allow any arbitrary text format in the diff highlighter, not just foreground colors
Previously, the diff highlighter maintained a "myColor" variable, that
was used to set the foreground color of the highlighted region based on
the parsing of the diff.

This patch changes myColor to myFormat; and (in principle) allows any
arbitrary formatting changes to be made to the highlighted block.
However, this patch does not make any additional changes; the diff
highlighting should appear exactly as it always did to the user.

Signed-off-by: Marco Costalba <mcostalba@gmail.com>