Add support for smart browsing
Let browse repository using only mouse wheel and/or
links in revision description pane in main view.

It is possible to directly jump to log, diff content,
up or down revision using links or mouse scrolling.

In the latter case the associated action is the
first link of the label.

Still some tweaks required, but more or less it seems
to work.

Signed-off-by: Marco Costalba <mcostalba@gmail.com>