Wheel browsing: fix false positive scroll event
Due to a logic error a bogus swicth command
was triggered when user scrolled all the view
starting with scrollbar at top/bottom screen.

This patch fixes the issue and also introdue a minimum
sroll step, used in case scroll bar is very near to
top/bottom position so that user is supposed to
want to start a swicthing command and not simply
scroll the content.

Finally labels position at startup should be ok now.

Signed-off-by: Marco Costalba <mcostalba@gmail.com>