Fix detection of a failing command
Check also if stderr is not empty for tagging a command
as failed. This is the way stable qgit has done for
ages without any issue.

Also tag as failed a 'canceled' command, i.e. a command
that we killed/terminated while running. This condition
should be already considered by process exit status, but
on Windows the flag doesn't seem so reliable.

This patch fixes a bug reported by Pavel Roskin, where
a failing 'stg fold' command was not reported to user.

Signed-off-by: Marco Costalba <mcostalba@gmail.com>