Fix a build problem on some platforms
Ismail Dönmez reported a build problem due
to -frepo option added to CXXFLAGS.

The cause of the issue is not clear but it seems,
googling a little bit, problems with -frepo are
rather common.

Just remove the option, the lose in bin size is in
any case small and compile times get much better.

Signed-off-by: Marco Costalba <mcostalba@gmail.com>