Fix a funny 'home run' behaviour
Annotate a file with a merge in history, then
select a revision below the merge point....surprise!

The selection will fall down till the first initial revision
in a fancy way.

We use two data container, one to store revisions row data
and one (a vector) to append the sha of loaded revisions.

Among the revision data there is the 'index', used as helper
in many places and set to be equal to the arrival order: if a
revision is the ten-th loaded its index will be 10.

To calculate the index is used, at revision creation
during laoding, the number of revisions already stored in
row data container. But because now this number could be
different from the number of 'arrived unique' revisions this
leads to a bug in case of merges where the secondary parent
are stored as row data, but are not added to sha vector to
not broke list view display code.

The fix is to use the sha vector count instead of row data
revisions count.