Further speedup filtering
Plug QAbstractProxyModel to the model only when
filtering is needed, not at startup and not when
only rows highlight filter is requested.

This fixes a little performance regression in normal case
(no filtering) and also speeds up highlight filtering
to be almost zero time.

When filter is released QAbstractProxyModel is unplugged
again, so that filter releasing is now immediate too.

Signed-off-by: Marco Costalba <mcostalba@gmail.com>