Add support for following renames in annotations
File history annotation is now able to detect and
follow file renames.

Renaming detect is done by "git diff-tree -M", if a file
is detected as renamed the file history of the old name
is retrived and appened to the new. This is done for any
following rename.

File renaming following is compatible with non
linear (graph) history.

Signed-off-by: Marco Costalba <mcostalba@gmail.com>