Speed up patch content viewer
Time to append text in QTextEdit is not related
to the quantity of text to append, but to the quantity
of text at the end of append. So that it takes more or
less the same time to add a big amount of text to an
empty QTextEdit or add a small text to a QTextEdit with an
already big content.

This patch tries to improve speed taking advantage
of this result.

Instead of append and show each new chunk of diff
data as soon as it is arrived from git, show only the
first one to give some feedback to the user, then
wait all patch content has been loaded and show it
at the end.

Also do not show the first chunk if patch viewer is
not visible, this is a common case of browsing with
bottom left pane set to revision description.

A downside of this patch is that in case of big merges
with "all files" flag set or in case of very big patch content
only the first few lines are shown and user has to wait
until the end of the whole loading to see the remaining
content. This could mean a couple of seconds in the
worst cases.

Signed-off-by: Marco Costalba <mcostalba@gmail.com>