Rename following: detect renaming starting from an old name
If user asks for annotation of a file that has been renamed
in the most recent branch tree then a backtracking of file
renames is performed before calling file history loading.

With this patch it is possible to annotate with renaming
any file selected in tree view from any point in revision
history.

Signed-off-by: Marco Costalba <mcostalba@gmail.com>