Update FileHistory to understand more then one file name
Now that we follow renames, the concept of one path per
file history is braindead broken.

This patch removes all the references to file names in
annotation and file history code and stores all the
renamed paths in history.

This new information is still not used, but will be a
prerequisite for future patches.

Signed-off-by: Marco Costalba <mcostalba@gmail.com>