Fix a segfault annotating renamed file not in file list
If the file to annotate is not in selcted revision's
file list, as example because selected directly from
tree view, a segfault could occur because file sha
is not found due to renaming.

Use file names list introduced in previous patch to
try harder to find the file sha, traversing all the
list.

Signed-off-by: Marco Costalba <mcostalba@gmail.com>