Rename following: try harder to find a file if not seen
In some cases we need to use 'git rev-list --full-history'
to be sure to find a given path searching from the top of
the branches.

This is the case of an old name of a file that for some reasons
has been pruned by simplified history walking of git-rev-list
So before to give up try again with the slower but more
accurate --full-history.

With this patch it is possible to get correct file history
of builtin-config.c also if invoked from his initial revision
when the name was config-set.c

Signed-off-by: Marco Costalba <mcostalba@gmail.com>