Profile tree indexing
Profiling tree indexing that is done after revisions
are loaded shows that converting the containers from
a QMap to a QHash could lead to great speed up expecially
in case of repositories with many tags.

Indeed the most time consuming task is checking if a given
revision is tagged.

In case of the git repo this patch cut indexing time from
420ms to less then 200ms on my box.

BTW KCachegrind it's a great tool!

Signed-off-by: Marco Costalba <mcostalba@gmail.com>