Select first revision as soon as possible
As soon as first revisions are loaded select the
first one in list instead of waiting until the end of
loading.

This gives a good 'snappy' feeling to user, also if
absolute loading time it's a little bigger due to
getting the file list of first rev while still loading.

Signed-off-by: Marco Costalba <mcostalba@gmail.com>