Use an optimized hash function for sha values
Because sha values are already hashed we don't need the
general purpose qHash() defined in Qt libraries.

Unfortunatly it is not possible to overload the global one
supplied by Qt, so use a template specialization of the QHash
function that calls qHash() for the classes that we
know 'for sure' have an sha as hash key.

With this patch we gain a 2X performance increase in
core revision lookup function.

One downside is that we need to copy a little part of Qt
sources, the ones that actually call qHash(), to tailor
an overloaded function that calls our fast-hashing one.

Signed-off-by: Marco Costalba <mcostalba@gmail.com>