Rewrite annotation core to use a one pass algorithm
Instead of copy previous annotation and then insert/remove
lines according to diff content start with an empty
annotation and build the new one in one pass.

This will allow to avoid using a linked list to store
annotation lines because insert/remove in the middle is not
more needed.

Signed-off-by: Marco Costalba <mcostalba@gmail.com>