New '\0' detection code uncovered a crash bug
When following renames, in case renamed file is
identical to old one, just reamed, the diff is empty
and we fake a "similarity index 100%" pseudo patch
to let the wheel go on.

The bug is this pseudo patch is not '\n' terminated
causing a crash in new '\0' detection code.

This patch fixes a crash when annotating builtin-mailsplit.c
in git repository.

Signed-off-by: Marco Costalba <mcostalba@gmail.com>