Re-add '--log-size' code but disabled
This new code can handle the case of msg size
not available, foreseen in git specification.

So it come easy to integrate with normal case of
git log called without --log-size option.

The difference is just a couple of lines. The result it's
a much cleaner code, compatible also with standard git.

Signed-off-by: Marco Costalba <mcostalba@gmail.com>