Fix annotation when selected rev is an StGit unapplied patch
In this case we don't find a valid ancestor in file's history because
git-log doesn't know unapplied patches and qgit raises an ASSERT.

The fix is to fallback on the first in history list. A real solution
would be to get the history also from unapplied patches, but that's
another story!

Bug reported by Yann Dirson.

Signed-off-by: Marco Costalba <mcostalba@gmail.com>