Add (disabled) support for git-log --early-output
With --early-output option git log prints out firsts
revisions as soon as available, without reordering the
whole tree so to avoid user waiting in front of a blank
screen for some seconds in case of big archives or cold
cache.

When final correct order revisions are ready then git-log
re-outputs the whole stream.

With this patch support for handling a replayed stream is
added but disabled for now, waiting for newly --early-option
to hit git main tree.

Signed-off-by: Marco Costalba <mcostalba@gmail.com>