Use custom format for git-log to shrink used memory
Ask git log only for needed info, no more useless
object parents, committer data, etc. Only a part of
all the stuff we get with --pretty=raw is actually
needed, so tailor output with --pretty=format so
to avoid allocating useless memory.

This shrinks allocated memory of 30% (17MB on Linux tree)
and there is also a good speedup in big repos when we are
low on RAM.

Signed-off-by: Marco Costalba <mcostalba@gmail.com>