Fix a crash due to random bad data from 'git log'
Apparently QFile::read() sometimes returns a byte array
with a heading '0' value. This is bogus and of unknown cause.

Anyhow recent parsing changes unhide this behaviour, that
was hidden before, crashing the application when this occurs.

The behaviour is rare but can be reproduced with luck and
multiple refreshes (F5) on a big repo as the git tree itself.

The patch workarounds this bogus zero value byte.

Signed-off-by: Marco Costalba <mcostalba@gmail.com>