Convert revisions hash and vector to use a QLatin1String
This Qt class is a a thin wrapper around an ASCII/Latin-1
encoded string literal.

Main advantage is that it doesn't make a deep copy of data
when is initialized with a const char*, this allows to avoid
a call to QString::fromLatin1() for each loaded revision
because sha value is not copied from laoded git data but is
referenced instead.

Another advantage is that additional memory to store sha
values is avoided.

Note that a copy is involved when passing from a QString to
a QLatin1String in particular a copy occurs for any revision
lookup. Further patches will try to avoid this at least in
the fast paths.

Also implicit conversion (with hidden copy) from QString will
be removed in the future to better spot data copies.

Signed-off-by: Marco Costalba <mcostalba@gmail.com>