Convert reference map hash to use new ShaString
The main reason here is to avoid a costly conversion in
a couple of inner loop points inside indexTree().

This function is used to find near tag/branch information
for each revision and is compute intensive expecially on big
repos because a walk of the entire tree is necessary.

With this patch indexTree() times are cutted in half, about
160ms from 310ms to index full git tree in my slow laptop.

Signed-off-by: Marco Costalba <mcostalba@gmail.com>