Load file names from cache while waiting for git revisions
Instead of defer cache data load at the end of git stream
output do this just after launching 'git log'.

This allow to load cache data in almost masked time for big
repos when first revisions take a while to arrive and also has
the added benefit to avoid a 'git diff-tree' call to load
file names for the first arrived revision simply because
cache file is still to be read from disk.

The only little cons is that total load time is a bit higher
because cache file reading and decompressing needs some I/O
and CPU cycle stolen to the git stream.

Note that when refreshing (F5) cache file is not reread, so
F5 is a good indicator of "pure" data loading performance.

Signed-off-by: Marco Costalba <mcostalba@gmail.com>