Use a QByteArray to store loaded file names indices
A single QByteArray is used instead of two vectors because it's
much faster to load from disk when using a QDataStream.

Paths are splitted in dir and file name, first all the
dirs are listed then the file names to achieve a better
compression when saved to disk.

Because names are loaded incrementally in temp variables
before to populate the QByteArray care should be taken
to avoid mixing partial results due to the fact that we
can have more simultaneous load of file names, so different
temporary variables grouped in different FileNamesLoader
struct instances should be used.

Unfortunatly this complicates a bit the code but advantages
are in much improved cache file loading times.

With this patch loading times are shrinked by a good 30%,
from 1093ms to 751ms on Linux tree on my slow laptop.

This is a noteworthy result given that file loading was
already heavily optimized.

Thanks to awsome kcachegrind profiler to spot the weak
point.

Signed-off-by: Marco Costalba <mcostalba@gmail.com>