Fix a data corruption bug in annotate
New ShaString code had a nasty bug that
sometimes could lead to data corruption.

Problem was a very well hidden stale pointer in
a ShaString, unfortunatly although memory was released
the effect of corruption was very rare because it was
used just after the memory release and 99.9% of cases
the memory was released but still valid, so to hide it.

This patch fixes annotation of builtin-fsck.c in git tree.

Signed-off-by: Marco Costalba <mcostalba@gmail.com>