Add a fancy progress bar while loading file names in background
If file names cache is not updated file names are loaded in background.

Add a progress bar so that new users are not scared by hard disk mistery
activity. File names loading is quite fast, but on big repos could last
also a minute the first time the repo is fully loaded.

Signed-off-by: Marco Costalba <mcostalba@gmail.com>