Fix broken 'stg pop' interface
On a stgit archive it is possible to pop a specified
applied patch from context menu (right click).

This feature was broken due to an upstream change in
stgit that changed pop behaviour.

The commit was "Allow patch ranges for the 'pop' command"
(d1368d811f2) of 6/12/2006, one year ago!

This patch fixes that.

Signed-off-by: Marco Costalba <mcostalba@gmail.com>