Teach qmake to check for correct Qt version
Checking the correct version (at least 4.3.0) is
a little bit tricky with qmake because string
comparison is not documented.

Anyway googling around I found this 'greaterThan'
function that seems to fit the bill.

With this patch an error message is shown at compilation
time in case Qt libraries are too old.