Fix 'filter by tree' switches to --all
When selecting a file and pressing magic wand tool
button then revisions are searched among all the
branches, not only the current one.

The patch does more then just fixing the branches,
indeed filter by tree is called with the same arguments
then the main view with added the selected paths.

Note that this is not exactly adding boolean "and" to the
selection, as example if the main view parameters are

HEAD -- tree.c

And you select tree-diff.c then the parameters used to
build the new filtered view will be

HEAD -- tree.c tree-diff.c

So you end up with more revisions then original, in this case.

Bug reported by Steffen Prohaska.

Signed-off-by: Marco Costalba <mcostalba@gmail.com>