Preserve last range select dialog settings
In case range dialog is requested again *on the same repo*
then previous settings are preserved.

Note that is not sensible to preserve settings across different
repositories, a solution could be to save as repo dependant
settings, but currently we don't have the framework for that,
so for now simply reset the values opening a new repository.

Usability bug reported by Steffen Prohaska.

Signed-off-by: Marco Costalba <mcostalba@gmail.com>