Restore default select of first tag in range dialog
In range dialog select the first tag that is not also
the current HEAD.

After the last patches the default become the first branch,
but this is 99% of cases unuseful because it is the
same of HEAD.

So user ends up almost always tweaking the range dialog
defaults before opening the repository.

With this patch default values are restored to more
sensible ones.

Signed-off-by: Marco Costalba <mcostalba@gmail.com>