Update and cleanup core commit code
It's time to update this very old part of code.

Problem is that it still relies on plumbing git commands
as 'git update-index' or 'git read-tree'.

This patch restructures the commit code to use newer and
officialy suggested 'git add', 'git rm' and 'git reset'
to update the index with the paths to commit.

As a side effect the code is much more clean now.

Signed-off-by: Marco Costalba <mcostalba@gmail.com>