Restore broken stg commit interface
Completly rewrite code to add a new patch or fold
working dir modifications in a stg patch.

It now uses 'git stash' to save modified files that
should not be added in the patch.

Unfortunatly stg is not very user friendly when you
need to create a patch with only a partial selection
of modified files, so the steps to perform the commit
are much more complex then in git case.

Signed-off-by: Marco Costalba <mcostalba@gmail.com>