Format patch: use selected patches as a range
Instead of pass patches one by one pass the patches as
an unique range from oldest to newest.

With this fix patch numbering is always correct, as a side
effect we loose the possibility to create patches for separate
ranges in one go, but this is a much less common scenario then
passing a consecutive range of revisions and in this latter
common case we were missing a correct patch numbering.

I found this new behaviour more useful for myself, so patch
it's acknowledged ;-)

Signed-off-by: Marco Costalba <mcostalba@gmail.com>