From Paul Gideon Dann: Fixing scrolling of diff view
Manual apply a patch from Paul:

Fixing scrolling of diff view when clicking on files in the right-hand pane.
Previously, the view would scroll vertically correctly, but it would also scroll horizontally to the end of the header, which is not desirable.
Scrolling was also occasionally incorrect when selecting the first file in the diff (the view would jump to a point below the start of the diff for that file)

Signed-off-by: Marco Costalba <mcostalba@gmail.com>