Revert "common.h: slightly reduced memory usage of RevFile"
It breaks binary compatibility of file names cache file saved on disk.

Status vector is saved on disk as a part of file name caching.
Because the reverted patch changes it's type from
QVector<int> to QVector<unsigned char> we have a
different binary data file saved on disk. This breaks
released qgit versions that instead rely on the fact
that status is a vector of int.

To properly handle this we should update also the file
cache version number so that previous versions of qgit
can detect binary compatibility is broken and recreate
the file as new.

But IMHO this seems like an overkill for this patch.

Signed-off-by: Marco Costalba <mcostalba@gmail.com>