Teach sha line edit to resolve references too
In the top right line edit, where the sha of the selected
revision is shown, we can paste a new sha, press return
and jump to that. It works also with abbreviated forms of
sha and in case the sha is too much abbreviated and there
is some ambiguity then all the sha that contain that substring
are shown.

From today it is possible to write also a reference name,
a branch name or anything else that can be resolved by
'git rev-parse' and we can jump directly to that pressing return.

Signed-off-by: Marco Costalba <mcostalba@gmail.com>