Create a separate submenu for tags
It is more logical and hierarchical correct not to begin listing tags in
the existing submenu. The remaining space of the main context menu is
halved for the branches and tags to be displayed before the submenu has
to be created.

A little downside is that now you can get 2 "More..." submenus for
branches and tags instead of only 1.

Signed-off-by: Markus Heidelberg <markus.heidelberg@web.de>
Signed-off-by: Marco Costalba <mcostalba@gmail.com>