Do not translate zero sha in log title
In revision log message pane, in bottom left of main
view the sha-ish strings written in the message are,
if possible, transalted to the corresponding revision
title.

This is handy for user that looks at uman readable titles
instead of revision's hex numbers.

This patch teach the transaltion logic to ignore the zero sha
string (40 consecutive '0' chars) because it has normally a
special meaning and we want to keep that instead of transalting
in a non sense like "nothing to commit" or "working dir changes".

Signed-off-by: Marco Costalba <mcostalba@gmail.com>