Fix "Save patch..." wrong revision order
When selecting more then one revision and calling
"Save patch..." the patches _could_ be created in
the wrong order.

It happens as example with Qt4.4 as reported by
Markus Heidelberg, while Qt4.3 doesn't have this problem.

This is due to how QItemSelectionModel::selectedRows()
is implemented across different Qt versions.

This patch always orders the selected revs from newest to
oldest to avoid relying on an undocumented feature and to
work correctly under any Qt version.

Signed-off-by: Marco Costalba <mcostalba@gmail.com>