Use correct log format for unapplied StGit patches
The commit 3a7141cb1db2a3911703915845b18a057c9b1cad added the committer
information to the git log format for the main tree, and changed the
parsing code to expect it, but neglected to also change the format for
unapplied StGit patches. This resulted in all the fields in an unapplied
patch being offset by one - the committer field would show the author, the
author field would show (an integer representation of) the date, the author
date would show garbage, and the first line of the commit message would be
lost.

This patch creates a common definition of the log format string to both
resolve the bug and help ensure that it is not repeated.

Signed-off-by: Zane Bitter <zbitter@redhat.com>