Check the result returned by the function QObject::connect()
The chk_connect macro is used to check the result returned by the function
QObject::connect() in debug mode, it looks like on assert() function.
However, in the release mode, unlike the assert() function - test expression
is not removed.
To use the macro in the code need include <assert.h>