Make Settings dialog accessible on macOS
The Settings dialog was inaccessible on macOS because Qt would detect
both the "Settings" and the "Setup Actions" menu items as having the
PreferencesRole, removing the former from the "Edit" menu and moving the
latter to the application menu as the Preferences menu item.

This moves "Settings" to the application menu and keeps "Setup Actions"
in the "Actions" menu.