Cleanup Qt4 compatibility, raise Qt minimum to 5.6.0
While Qt4 is unmaintained for several years and itself fails to build
on modern systems, qgit build with Qt4 was broken by commits b68c3e76
and 3f25d12a and no one seems to have noticed in ~two years.

Rather than fix it, it is probably time to clean it up altogether.

This commit does not fix qmake build system which seems to have more
issues, it should probably be dropped as well in favor of CMake.

Signed-off-by: Andreas Sturmlechner <asturm@gentoo.org>